pyproject.toml
[project]
name = "knowledge-graphs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.2",
    "pydatalog>=0.17.4",
    "pypinyin>=0.55.0",
    "python-dotenv>=1.1.1",
]
